+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # V1.1.0 - unreleased
+## Items added
+- Ritual Mask
+- Watcher Shield
 
 # v1.0.0
 This in the first public release!